Added HTTP server and client examples

+++ bitset.h
@@ -16,19 +16,19 @@ namespace co {
 class BitSet {
  public:
   // Allocate the first free bit.
-  int Allocate();
+  uint32_t Allocate();
 
   // Free a bit.
-  void Free(size_t bit);
+  void Free(uint32_t bit);
 
   // Set a bit.
-  void Set(size_t bit);
+  void Set(uint32_t bit);
 
   // Is the bitset empty (all bits clear)?
   bool IsEmpty() const;
 
   // Is the given bit set?
-  bool Contains(size_t bit) const;
+  bool Contains(uint32_t bit) const;
 
  private:
   // Note the use of explicit long long type here because
@@ -37,11 +37,11 @@ class BitSet {
   std::vector<long long> bits_;
 };
 
-inline int BitSet::Allocate() {
-  size_t start = 0;
+inline uint32_t BitSet::Allocate() {
+  uint32_t start = 0;
   for (;;) {
-    for (size_t i = start; i < bits_.size(); i++) {
-      int bit = ffsll(~bits_[i]);
+    for (uint32_t i = start; i < bits_.size(); i++) {
+      uint32_t bit = static_cast<uint32_t>(ffsll(~bits_[i]));
       if (bit != 0) {
         bits_[i] |= (1 << (bit - 1));
         return i * 64 + (bit - 1);
@@ -56,26 +56,26 @@ inline int BitSet::Allocate() {
   }
 }
 
-inline void BitSet::Free(size_t bit) {
-  size_t word = bit / 64;
+inline void BitSet::Free(uint32_t bit) {
+  uint32_t word = bit / 64;
   if (word < 0 || word >= bits_.size()) {
     return;
   }
-  int b = bit % 64;
+  uint32_t b = bit % 64;
   bits_[word] &= ~(1 << b);
 }
 
-inline void BitSet::Set(size_t bit) {
-  size_t word = bit / 64;
+inline void BitSet::Set(uint32_t bit) {
+  uint32_t word = bit / 64;
   if (word < 0 || word >= bits_.size()) {
     return;
   }
-  int b = bit % 64;
+  uint32_t b = bit % 64;
   bits_[word] |= (1 << b);
 }
 
 inline bool BitSet::IsEmpty() const {
-  for (size_t i = 0; i < bits_.size(); i++) {
+  for (uint32_t i = 0; i < bits_.size(); i++) {
     if (bits_[i] != 0) {
       return false;
     }
@@ -83,12 +83,12 @@ inline bool BitSet::IsEmpty() const {
   return true;
 }
 
-inline bool BitSet::Contains(size_t bit) const {
-  size_t word = bit / 64;
+inline bool BitSet::Contains(uint32_t bit) const {
+  uint32_t word = bit / 64;
   if (word < 0 || word >= bits_.size()) {
     return false;
   }
-  int b = bit % 64;
+  uint32_t b = bit % 64;
   return (bits_[word] & (1 << b)) != 0;
 }
 }  // namespace co
